Create README.md
I created a table for baby milestones from 0 to 24 months. Then inserted data by milestone and age. I noticed I had created a mistake for milestone 4. The milestone was listed for 8.54 months and I corrected it to 8.5 months. I then corrected this in the table and deleted 'painting' as a milestone due to no longer wanting this piece of data.